+++ tests/Unit/Framework/Container/ContainerTest.php
@@ -258,4 +258,27 @@ final class ContainerTest extends TestCase
         $this->expectExceptionObject(ContainerException::serviceFrozen('service_name'));
         $this->container->set('service_name', static fn () => new ArrayObject([3]));
     }
+
+    public function test_protect_service_is_not_resolved(): void
+    {
+        $service = static fn () => 'value';
+        $this->container->set('service_name', $this->container->protect($service));
+
+        self::assertSame($service, $this->container->get('service_name'));
+    }
+
+    public function test_protect_service_cannot_be_extended(): void
+    {
+        $this->container->set(
+            'service_name',
+            $this->container->protect(new ArrayObject([1, 2]))
+        );
+
+        $this->expectExceptionObject(ContainerException::serviceProtected('service_name'));
+
+        $this->container->extend(
+            'service_name',
+            static fn (ArrayObject $arrayObject) => $arrayObject
+        );
+    }
 }